Add a pass to unfreeze mutable global tensors and restore them as `arith.ConstOp`->`tf.VarHandleOp` -> `tf.AssignVariableOp` in the session_initializer.
This pass is added to the TFLite converter to support mutable global tensors in TFLite. The pass is added after the `OptimizeGlobalTensorsPass` and `FreezeGlobalTensorsPass` passes. The pass first finds all the mutable global tensors in the module. Then, it creates a new `tf.VarHandleOp` for each mutable global tensor and assigns the const op's value to the `tf.VarHandleOp`. Finally, it removes the references to the mutable global tensors from the module.

PiperOrigin-RevId: 649899938