Fix how we read the `xla_use_shardonnay` flag when delaying the use of tuples.
Before we were reading it based on flags. But the shardy debug option isn't always set based on flags. So pass it in from the `CompileOptions`.

Don't need to propagate this through everywhere, and it's best to use false in most places where API changes are needed, as this fix is temporary until Shardy is the first pass in the XLA pipeline. Delaying using tuple arguments in the Shardy code path is only because Shardy is in the middle of the XLA pipeline. This will not be the case in the future.

PiperOrigin-RevId: 651768186